Moves routes to their own `routes.py` module
